Working at making it working.

+++ minisound_ffi/src/include/sound_data/encoded_sound_data.h
@@ -1,6 +1,8 @@
 #ifndef ENCODED_SOUND_DATA_H
 #define ENCODED_SOUND_DATA_H
 
+#include <stdint.h>
+
 #include "../../external/result/result.h"
 #include "sound_data.h"
 
@@ -9,7 +11,7 @@ typedef struct EncodedSoundData EncodedSoundData;
 EncodedSoundData *encoded_sound_data_alloc(void);
 Result encoded_sound_data_init(
     EncodedSoundData *const self,
-    float const *const data,
+    uint8_t const *const data,
     size_t const data_size
 );
 void encoded_sound_data_uninit(EncodedSoundData *const self);
